+++ test/helpers/lint.js
@@ -4,11 +4,12 @@ const path = require('node:path');
 
 const pluginPath = path.join(__dirname, '..', '..', 'src', 'index.js');
 
-async function lintCss({ code, fix = false, rules }) {
+async function lintCss({ code, customSyntax, fix = false, rules }) {
   const { default: stylelint } = await import('stylelint');
   const lintResult = await stylelint.lint({
     code,
     config: {
+      ...(customSyntax ? { customSyntax } : {}),
       plugins: [pluginPath],
       rules,
     },